+++ main.ts
@@ -4,8 +4,8 @@ import {
   PluginSettings,
   resolveBlockSettings,
 } from "./src/settings/schema";
-import { renderExplorerBlock } from "./src/plugin/explorer";
-import { parseSettings } from "./src/block-settings";
+import { renderExplorerBlock } from "./src/explorer";
+import { parseSettings } from "./src/settings/block-parser";
 import { ExplorerSettingsTab } from "./src/ui/settings-tab";
 
 export default class ExplorerPlugin extends Plugin {